refactor: extract 5 multi-value read panels from PanelViews.tsx
Batch 3 of the per-panel split for issue #144: worksheets,
assigned_schools, student_progress, attendance, teachers — panels
needing 2-3 of students/schools/teachersList/reportsList/worksheetsList
from usePanelData.

- New panels/WorksheetsPanel.tsx (reportsList + worksheetsList),
  AssignedSchoolsPanel.tsx (schools + students), StudentProgressPanel.tsx
  (students only), AttendancePanel.tsx (students + reportsList),
  TeachersPanel.tsx (schools + teachersList + currentUser for the
  isBlockAdmin check) — each moved verbatim, taking only the props
  they use.
- teachers' extra role guard (SCHOOL/BLOCK_ADMIN only) stays in
  PanelViews.tsx's router line, not inside the component.
- PanelViews.tsx's 5 if-blocks replaced with one-line router calls.
- Unused-import cleanup (same audit method as PR3): removed
  CheckCircle2, XCircle, ClipboardList, GraduationCap (all confirmed
  dead - their only usages were in the panels moved here) and
  EvaluationReport (confirmed dead - only usage was worksheets' own
  type annotation, which now lives in WorksheetsPanel.tsx itself).
  Left Worksheet (the type) alone - confirmed it was never a real type
  usage in PanelViews.tsx even pre-refactor, just comment/string
  false-positive matches.

Also found while verifying: assigned_schools, student_progress, and
attendance are not wired into any role's nav in Layout.tsx - genuinely
unreachable in the live app today (pre-existing, unrelated to this
PR - not a crash bug like #245, just dead UI paths). Confirmed via
grep against Layout.tsx's role->nav-item mapping, not touched here.

Verified: tsc --noEmit clean, vite build succeeds, byte-diff of every
moved block. Live click-through covered every panel actually reachable
via nav: worksheets (Teacher, honest 0-state), teachers (School
Principal, isBlockAdmin=false description text confirmed correct).
assigned_schools/student_progress/attendance verified via tsc+build+
byte-diff only, since no nav path reaches them in the current app.
